+++ src/index.html
@@ -19,9 +19,9 @@
                 <label for="loginUsername">Username:</label>
                 <input id="loginUsername" name="loginUsername" type="text" required minlength="3" maxlength="15">
                 <label for= "loginPassword">Password:</label>
-                <input id="loginPassword" name="loginPassword" type="password" required minlength="8" maxlength="15">
+                <input id="loginPassword" name="loginPassword" type="password" required minlength="3" maxlength="15">
             </form>
-            <button id="login--button">Submit</button>
+            <button id="login__button">Submit</button>
         </section>
     <!-- Added Primary Banner Breakdown-->
     <header class= "Header">
