+++ src/scripts/AuthenticationProvider/authenticationClickEvents.js
@@ -1,4 +1,4 @@
-import API from "./authenicationProvider.js"
+import API from "./authenticationProvider.js"
 
 const loginSubmitButton = document.querySelector("#login__button")
 
@@ -8,10 +8,10 @@ const loginSubmitButton = document.querySelector("#login__button")
             console.log("I click first", event.target.id)
             const userLogin = document.querySelector("#loginUsername").value
             const passwordLogin = document.querySelector("#loginPassword").value
-            API.getSingleUser()
+            API.getSingleUser(userLogin)
             .then((loginArray) => {
                 // clearInputs()
-                console.log(loginId)
+                console.log(userLogin)
                 console.log(loginArray)
                 console.log("click last")
             })
